NotificationsSider components separation
