+++ src/components/NotificationsSider/NotificationsSider.module.scss
